requirements.txt
pandas==1.5.3
sklearn==1.3.0
xgboost==1.7.3
hydra-core==1.3.2
matplotlib==3.6.2
gradio==3.40.1

# local package
-e.

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
